feat: adicionando validação caso id não exista

+++ src/route.js
@@ -48,6 +48,15 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params
+
+      const task = database.select('tasks').find(row => row.id === id)
+
+      if(!task) {
+        return res.writeHead(404).end(JSON.stringify({
+          message: 'task not found'
+        }))
+      }
+
       database.delete('tasks', id)
 
       return res.writeHead(204).end(JSON.stringify({
